+++ CHANGELOG.md
@@ -17,6 +17,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased](https://github.com/o1-labs/o1js/compare/b04520d...HEAD)
 
+### Added
+
+- Support secp256r1 in elliptic curve and ECDSA gadgets https://github.com/o1-labs/o1js/pull/1885
+
 ## [2.0.0](https://github.com/o1-labs/o1js/compare/7e9394...b04520d)
 
 ### Breaking Changes